tools/mlx-lm/requirements.txt
mlx-lm
huggingface_hub
python-dotenv
